aks-eu-tap-8

+++ env.sh
@@ -5,7 +5,7 @@ export GIT_SSH_PRIVATE_KEY=$(cat ~/dotconfig/.ssh/id_rsa)
 export GIT_KNOWN_HOSTS=$(ssh-keyscan github.com)
 export SOPS_AGE_KEY=$(cat ${2})
 export SOPS_AGE_KEY_FILE=${2}
-export SOPS_AGE_RECIPIENTS=`cat ${SOPS_AGE_KEY_FILE} | grep "# public key: " | sed 's/# public key: //'`
+export SOPS_AGE_RECIPIENTS=$(cat ${SOPS_AGE_KEY_FILE} | grep "# public key: " | sed 's/# public key: //')
 export TAP_PKGR_REPO=${INSTALL_REGISTRY_HOSTNAME}/tanzu-application-platform/tap-packages
 export CLUSTER_NAME=${3}
 
@@ -19,7 +19,8 @@ echo "GIT_KNOWN_HOSTS           \n${GIT_KNOWN_HOSTS}\n/GIT_KNOWN_HOSTS"
 echo "GIT_SSH_PRIVATE_KEY       \n${GIT_SSH_PRIVATE_KEY}\n/GIT_SSH_PRIVATE_KEY"
 
 SENSITIVE_FILE_NAME=clusters/${CLUSTER_NAME}/tanzu-sync/app/values/tanzu-sync-values-sensitive.yaml
-sensitive_tanzu_sync_values=$(cat << EOF
+sensitive_tanzu_sync_values=$(
+  cat <<EOF
 ---
 secrets:
   sops:
@@ -38,6 +39,8 @@ $(echo "$GIT_KNOWN_HOSTS" | awk '{printf "          %s\n", $0}')
 EOF
 )
 
-echo "${sensitive_tanzu_sync_values}" > ${SENSITIVE_FILE_NAME}
-sops --encrypt  ${SENSITIVE_FILE_NAME} >  clusters/${CLUSTER_NAME}/tanzu-sync/app/sensitive-values/tanzu-sync-values.sops.yaml 
-rm ${SENSITIVE_FILE_NAME}
+if [ -d "$FILE" ]; then
+  echo "${sensitive_tanzu_sync_values}" >${SENSITIVE_FILE_NAME}
+  sops --encrypt ${SENSITIVE_FILE_NAME} >clusters/${CLUSTER_NAME}/tanzu-sync/app/sensitive-values/tanzu-sync-values.sops.yaml
+  rm ${SENSITIVE_FILE_NAME}
+fi
